OCR/requirements.txt
google-cloud-vision
google-cloud-documentai
pdfplumber
tabulate
